+++ copy-text/index.commonjs.js
@@ -1,4 +1,0 @@
-'use strict'
-
-exports = module.exports = require('../commonjs/ReactTimeAgoCopyText').default
-exports['default'] = require('../commonjs/ReactTimeAgoCopyText').default