+++ no-tooltip.js
@@ -1,4 +1,0 @@
-var ReactTimeAgo = require('./commonjs/ReactTimeAgo').default
-
-exports = module.exports = ReactTimeAgo
-exports['default']       = ReactTimeAgo