+++ website/lib/locales.en.js
@@ -1,0 +1,417 @@
+var long = {
+  "year": {
+    "future": {
+      "one": "in a year",
+      "other": "in {0} years"
+    },
+    "past": {
+      "one": "a year ago",
+      "other": "{0} years ago"
+    }
+  },
+  "half-year": {
+    "future": {
+      "other": "in half a year"
+    },
+    "past": {
+      "other": "half a year ago"
+    }
+  },
+  "month": {
+    "future": {
+      "one": "in a month",
+      "other": "in {0} months"
+    },
+    "past": {
+      "one": "a month ago",
+      "other": "{0} months ago"
+    }
+  },
+  "week": {
+    "future": {
+      "one": "in a week",
+      "other": "in {0} weeks"
+    },
+    "past": {
+      "one": "a week ago",
+      "other": "{0} weeks ago"
+    }
+  },
+  "day": {
+    "future": {
+      "one": "in a day",
+      "other": "in {0} days"
+    },
+    "past": {
+      "one": "a day ago",
+      "other": "{0} days ago"
+    }
+  },
+  "hour": {
+    "future": {
+      "one": "in an hour",
+      "other": "in {0} hours"
+    },
+    "past": {
+      "one": "an hour ago",
+      "other": "{0} hours ago"
+    }
+  },
+  "half-hour": {
+    "future": {
+      "other": "in half an hour"
+    },
+    "past": {
+      "other": "half an hour ago"
+    }
+  },
+  "minute": {
+    "future": {
+      "one": "in a minute",
+      "other": "in {0} minutes"
+    },
+    "past": {
+      "one": "a minute ago",
+      "other": "{0} minutes ago"
+    }
+  },
+  "second": {
+    "future": {
+      "one": "in a second",
+      "other": "in {0} seconds"
+    },
+    "past": {
+      "one": "a second ago",
+      "other": "{0} seconds ago"
+    }
+  },
+  "now": {
+    "future": "in a moment",
+    "past": "just now"
+  }
+}
+
+var long_time = {
+  "year": {
+    "future": {
+      "one": "{0} year",
+      "other": "{0} years"
+    },
+    "past": {
+      "one": "{0} year",
+      "other": "{0} years"
+    }
+  },
+  "half-year": {
+    "future": {
+      "other": "half a year"
+    },
+    "past": {
+      "other": "half a year"
+    }
+  },
+  "month": {
+    "future": {
+      "one": "{0} month",
+      "other": "{0} months"
+    },
+    "past": {
+      "one": "{0} month",
+      "other": "{0} months"
+    }
+  },
+  "week": {
+    "future": {
+      "one": "{0} week",
+      "other": "{0} weeks"
+    },
+    "past": {
+      "one": "{0} week",
+      "other": "{0} weeks"
+    }
+  },
+  "day": {
+    "future": {
+      "one": "{0} day",
+      "other": "{0} days"
+    },
+    "past": {
+      "one": "{0} day",
+      "other": "{0} days"
+    }
+  },
+  "hour": {
+    "future": {
+      "one": "{0} hour",
+      "other": "{0} hours"
+    },
+    "past": {
+      "one": "{0} hour",
+      "other": "{0} hours"
+    }
+  },
+  "half-hour": {
+    "future": {
+      "other": "half an hour"
+    },
+    "past": {
+      "other": "half an hour"
+    }
+  },
+  "minute": {
+    "future": {
+      "one": "{0} minute",
+      "other": "{0} minutes"
+    },
+    "past": {
+      "one": "{0} minute",
+      "other": "{0} minutes"
+    }
+  },
+  "second": {
+    "future": {
+      "one": "{0} second",
+      "other": "{0} seconds"
+    },
+    "past": {
+      "one": "{0} second",
+      "other": "{0} seconds"
+    }
+  },
+  "now": {
+    "future": "a moment",
+    "past": "just now"
+  }
+}
+
+var short = {
+  "year": {
+    "future": {
+      "one": "in {0} yr.",
+      "other": "in {0} yr."
+    },
+    "past": {
+      "one": "{0} yr. ago",
+      "other": "{0} yr. ago"
+    }
+  },
+  "month": {
+    "future": {
+      "one": "in {0} mo.",
+      "other": "in {0} mo."
+    },
+    "past": {
+      "one": "{0} mo. ago",
+      "other": "{0} mo. ago"
+    }
+  },
+  "week": {
+    "future": {
+      "one": "in {0} wk.",
+      "other": "in {0} wk."
+    },
+    "past": {
+      "one": "{0} wk. ago",
+      "other": "{0} wk. ago"
+    }
+  },
+  "day": {
+    "previous": "a day ago",
+    "next": "in a day",
+    "future": {
+      "one": "in {0} day",
+      "other": "in {0} days"
+    },
+    "past": {
+      "one": "{0} day ago",
+      "other": "{0} days ago"
+    }
+  },
+  "hour": {
+    "future": {
+      "one": "in {0} hr.",
+      "other": "in {0} hr."
+    },
+    "past": {
+      "one": "{0} hr. ago",
+      "other": "{0} hr. ago"
+    }
+  },
+  "minute": {
+    "future": {
+      "one": "in {0} min.",
+      "other": "in {0} min."
+    },
+    "past": {
+      "one": "{0} min. ago",
+      "other": "{0} min. ago"
+    }
+  },
+  "second": {
+    "past": {
+      "one": "{0} sec. ago",
+      "other": "{0} sec. ago"
+    },
+    "future": {
+      "one": "in {0} sec.",
+      "other": "in {0} sec."
+    }
+  },
+  "now": {
+    "past": "now",
+    "future": "now"
+  }
+}
+
+var short_time = {
+  "year": {
+    "future": {
+      "one": "{0} yr.",
+      "other": "{0} yr."
+    },
+    "past": {
+      "one": "{0} yr.",
+      "other": "{0} yr."
+    }
+  },
+  "month": {
+    "future": {
+      "one": "{0} mo.",
+      "other": "{0} mo."
+    },
+    "past": {
+      "one": "{0} mo.",
+      "other": "{0} mo."
+    }
+  },
+  "week": {
+    "future": {
+      "one": "{0} wk.",
+      "other": "{0} wk."
+    },
+    "past": {
+      "one": "{0} wk.",
+      "other": "{0} wk."
+    }
+  },
+  "day": {
+    "future": {
+      "one": "{0} day",
+      "other": "{0} days"
+    },
+    "past": {
+      "one": "{0} day",
+      "other": "{0} days"
+    }
+  },
+  "hour": {
+    "future": {
+      "one": "{0} hr.",
+      "other": "{0} hr."
+    },
+    "past": {
+      "one": "{0} hr.",
+      "other": "{0} hr."
+    }
+  },
+  "minute": {
+    "future": {
+      "one": "{0} min.",
+      "other": "{0} min."
+    },
+    "past": {
+      "one": "{0} min.",
+      "other": "{0} min."
+    }
+  },
+  "second": {
+    "past": {
+      "one": "{0} sec.",
+      "other": "{0} sec."
+    },
+    "future": {
+      "one": "{0} sec.",
+      "other": "{0} sec."
+    }
+  },
+  "now": {
+    "past": "now",
+    "future": "now"
+  }
+}
+
+var tiny = {
+  "year": {
+    "future": {
+      "other": "{0}yr"
+    },
+    "past": {
+      "other": "{0}yr"
+    }
+  },
+  "month": {
+    "future": {
+      "other": "{0}mo"
+    },
+    "past": {
+      "other": "{0}mo"
+    }
+  },
+  "week": {
+    "future": {
+      "other": "{0}wk"
+    },
+    "past": {
+      "other": "{0}wk"
+    }
+  },
+  "day": {
+    "future": {
+      "other": "{0}d"
+    },
+    "past": {
+      "other": "{0}d"
+    }
+  },
+  "hour": {
+    "future": {
+      "other": "{0}h"
+    },
+    "past": {
+      "other": "{0}h"
+    }
+  },
+  "minute": {
+    "future": {
+      "other": "{0}m"
+    },
+    "past": {
+      "other": "{0}m"
+    }
+  },
+  "second": {
+    "past": {
+      "other": "{0}s"
+    },
+    "future": {
+      "other": "{0}s"
+    }
+  },
+  "now": {
+    "past": "",
+    "future": ""
+  }
+}
+
+window['javascript-time-ago/locales/en'] =
+{
+  locale : 'en',
+
+  tiny       : tiny,
+  short      : short,
+  short_time : short_time,
+  long       : long,
+  long_time  : long_time,
+
+  quantify : function(n){var r=!String(n).split(".")[1];return 1==n&&r?"one":"other"}
+}